+++ jest.config.js
@@ -6,5 +6,11 @@ module.exports = {
   verbose: true,
   bail: true,
   testMatch: [files],
-  testEnvironment: process.env.TEST_ENV || 'jsdom'
+  testEnvironment: process.env.TEST_ENV || 'jsdom',
+  transform: {
+      "^.+\\.js$": "babel-jest",
+  },
+  transformIgnorePatterns: [
+    "node_modules/(?!promise-polyfill|whatwg-fetch)/"
+  ]
 };