+++ CHANGELOG.md
@@ -6,6 +6,16 @@
 **BREAKING:**
 **CHANGE:**
 -->
+# Fetch & Promises
+<a name="2.0.2"></a>
+
+**New:**
+* Fetch API is now the default method of communication.
+* client.recordEvent returns a Promise
+
+**FIXED:**
+* auto-tracking listeners will wait more than 500ms if necessary
+
 <a name="2.0.1"></a>
 
 **FIXED:**